feat: Update SerialMultiplexer class to use int instead of uint8_t for channel selection

+++ include/SerialMultiplexer.h
@@ -1,10 +1,11 @@
 #ifndef SERIAL_MULTIPLEXER_H
 #define SERIAL_MULTIPLEXER_H
+#include <Arduino.h>
 
 class SerialMultiplexer
 {
 public:
-    void tcaSelect(uint8_t i); // Function to select TCA9548A channel
+    void tcaSelect(int i);
 
 private:
     const int TCAADDR = 0x70; // I2C address of the TCA9548A
